Moved all the safeXXX methods into typesafe.cpp
