+++ src/Makefile
@@ -1,8 +1,8 @@
 CXX=g++
-CXXFLAGS=-O3 -std=c++11 -Wall
+CXXFLAGS= -std=c++11 -Wall
 #CFLAGS = -O3 -std=c++14 -stdlib=libc++ -Wall -Wno-nested-anon-types -Wno-gnu-anonymous-struct -Wunused
 HFILES	= madc.h datadef.h tokens.h datatokens.h
-OFILES	= lexer.o parser.o compiler.o madc.o
+OFILES	= lexer.o parser.o typesafe.o compiler.o madc.o
 INCDIR	= ../include
 BINDIR	= ../bin
 LIBDIR	= ../lib
